Water Tracker: Updated the water tracker visual along with cleaning up the front end. Did not the changes to the visual based on input yet.
Dashboard: Update the visual of dashboard along with cleaning up the whole dashboard. Dashboard now has a calorie visual and a macro visual. Implemented the ability to swipe to see different visuals on the dashboard. Reformatted the layout because of multiple nested vertical layouts/textboxes. Changed the fun fact of the day to a randomized fun fact. ( Will possibly implement a randomize fun fact button or keep it as a random fact for the day). Added a fun fact bank.